pyproject.toml
[project]
name = "ntgram"
version = "0.1.0"
description = "Python MTProto-compatible backend with service-oriented architecture."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvloop>=0.20.0",
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "protobuf>=5.27.0",
    "redis>=5.0.0",
    "asyncpg>=0.30.0",
    "cryptography>=44.0.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["src/ntgram/gen"]
